Remove stain caused by faux mono-repo
Some gits established a "mono-repo" that had an external dependency on a Go
package that was itself a disparate, private git repository, hence a terrible
hack had to be used to satiate `go get` (but breaks cargo!)  in addition to
forcing users to remove their ssh passwords.

Since we're not puttering around with such gits anymore, remove the kludge
out of our git configuration.