Add beat-and-measuer counter overlay script for Reaper video effect plugin
Displays the BPM, beat count (1 thru 8), and incrementing measure count
as a video overlay for a project in Reaper. Add a pickup measure in
project settings, and also set the measure length to 8/4.

This is meant for creating videos to teach people complex choreo,
where they can pause, stop, or slow down and always know where they are
from the overlay.

This "code" in my configurations repository because I have no idea where
else to put it at the moment and it was non-trivial to write.